+++ src/device/device_default.cpp
@@ -373,11 +373,6 @@ namespace hw {
             return rct::genCommitmentMask(amount_key);
         }
 
-        bool  device_default::ecdhEncode(rct::ecdhTuple & unmasked, const rct::key & sharedSec, bool short_amount) {
-            rct::ecdhEncode(unmasked, sharedSec, short_amount);
-            return true;
-        }
-
         bool  device_default::ecdhDecode(rct::ecdhTuple & masked, const rct::key & sharedSec, bool short_amount) {
             rct::ecdhDecode(masked, sharedSec, short_amount);
             return true;